melhorias na tela de try out da cifra de césar
